Move API services and guards to the shared module, auth service to the authentication module
