+++ app/app.routing.ts
@@ -1,11 +1,11 @@
 import { ModuleWithProviders } from '@angular/core';
 import { Routes, RouterModule }   from '@angular/router';
 
-import { AuthGuard } from './core/guards/auth-guard';
-import { AuthApiService } from './core/api/auth-api.service';
+import { AuthGuard } from './shared/guards/auth-guard';
+import { AuthApiService } from './shared/api/auth-api.service';
 import { AuthService } from './authentication/auth.service';
-import { CmsApiService } from './core/api/cms-api.service';
-import { SupervisorGuard } from './core/guards/supervisor-guard';
+import { CmsApiService } from './shared/api/cms-api.service';
+import { SupervisorGuard } from './shared/guards/supervisor-guard';
 import { UserService } from './users/user.service';
 
 const appRoutes: Routes = [
