+++ app/topics/content/content.component.ts
@@ -6,7 +6,7 @@ import { TranslateService } from 'ng2-translate';
 import { TopicService } from '../shared/topic.service';
 import { UserService } from '../../users/user.service';
 import { User } from '../../users/user.model';
-import { OOApiService } from '../../core/api/oo-api.service';
+import { OOApiService } from '../../shared/api/oo-api.service';
 import { ConfigService } from '../../config.service';
 
 declare let window: {
