+++ app/topics/topics.routing.ts
@@ -3,7 +3,7 @@ import { Routes, RouterModule } from '@angular/router';
 
 import { AddNewSubtopicComponent } from './topic-management/add-new-subtopic/add-new-subtopic.component';
 import { AllTopicsComponent } from './all-topics/all-topics.component';
-import { AuthGuard } from '../core/guards/auth-guard';
+import { AuthGuard } from '../shared/guards/auth-guard';
 import { ContentComponent } from './content/content.component';
 import { DeleteTopicComponent } from './topic-management/delete-topic/delete-topic.component';
 import { EditTopicComponent } from './topic-management/edit-topic/edit-topic.component';
@@ -11,7 +11,7 @@ import { ManageAttachmentsComponent } from './topic-management/manage-attachment
 import { MyTopicsComponent } from './my-topics-list/my-topics-list.component';
 import { NewTopicComponent } from './topic-management/new-topic/new-topic.component';
 import { ShowTopicComponent } from './show-topic/show-topic.component';
-import { SupervisorGuard } from '../core/guards/supervisor-guard';
+import { SupervisorGuard } from '../shared/guards/supervisor-guard';
 import { UploadDocumentComponent } from './topic-management/upload-document/upload-document.component';
 
 const topicRoutes: Routes = [
